refactor: move folder components for shadcn inside src and own components inside app, make looping for dummy data on hompage in page.tsx
