+++ amimum/src/app/components/ProductionCard/index.tsx
@@ -1,11 +1,9 @@
 import Image from "next/image";
-import jamuJago from "../../../public/assets/images/jamu_jago_1.png";
-
 const ProductionCard = () => {
   return (
     <div className="bg-white rounded-lg h-36 flex flex-col items-center justify-center w-24 shadow-md">
       <div className="bg-gray-100 rounded-lg px-3 py-2 flex flex-col justify-center items-center">
-        <Image src={jamuJago} alt="promo" width={50} height={50} />
+        <Image src="/jamu_jago_1.png" alt="promo" width={50} height={50} />
       </div>
 
       <div className="flex flex-col justify-center items-center mt-2 gap-1">
