+++ amimum/src/app/components/index.tsx
@@ -1,0 +1,7 @@
+export {default as AccordionExpandDefault} from "./AccordionExpandDefault"
+export {default as Carousel} from "./Carousel"
+export {default as Navbar} from "./Navbar"
+export {default as ProductionCard} from "./ProductionCard"
+export {default as PromoCard} from "./PromoCard"
+export {default as Search} from "./Search"
+export {default as Tag} from "./Tag"
